+++ gulpfile.js
@@ -2,6 +2,7 @@ var gulp = require('gulp')
 var rename = require('gulp-rename')
 var jshint = require('gulp-jshint')
 var browserify = require('gulp-browserify')
+var mocha = require('gulp-mocha')
 
 var paths = {
   entry: 'public/js/main.js',
@@ -16,7 +17,8 @@ var paths = {
       'fixture/**/*.js'
     ]
   },
-  dist: 'public/dist'
+  dist: 'public/dist',
+  test: 'test/**/*.spec.js'
 }
 
 gulp.task('build', function () {
@@ -53,4 +55,10 @@ gulp.task('initdb', function () {
   setTimeout(function () {
     process.exit()
   }, 2000)
+})
+
+gulp.task('test', function () {
+  gulp
+    .src(paths.test)
+    .pipe(mocha())
 })
