+++ package.json
@@ -34,7 +34,12 @@
     "gulp-rename": "^1.2.0",
     "browserify": "^4.2.3",
     "browserify-shim": "^3.6.0",
-    "gulp-jshint": "^1.8.4"
+    "gulp-jshint": "^1.8.4",
+    "gulp-mocha": "~1.0.0",
+    "expect": "~0.1.1",
+    "should": "~4.0.4",
+    "mocha": "~1.21.4",
+    "supertest": "~0.13.0"
   },
   "bundleDependencies": [],
   "private": true,
